+++ test/send.js
@@ -635,20 +635,6 @@ describe('send(file).pipe(res)', function(){
       .expect('Cache-Control', 'public, max-age=31536000', done)
     })
   })
-
-  describe('.root()', function(){
-    it('should set root', function(done){
-      var app = http.createServer(function(req, res){
-        send(req, req.url)
-        .root(__dirname + '/fixtures')
-        .pipe(res);
-      });
-
-      request(app)
-      .get('/pets/../name.txt')
-      .expect(200, 'tobi', done)
-    })
-  })
 })
 
 describe('send(file, options)', function(){